+++ src/connect/ncbi_sendmail.c
@@ -285,7 +285,7 @@ static void x_Sendmail_InitEnv(void)
         tmo = 120.0/*2 min*/;
     }
     if (!ConnNetInfo_GetValue(0, "MX_PORT", buf, sizeof(buf), 0)
-        ||  !(port = atoi(buf))  ||  port > 65535) {
+        ||  !(port = (unsigned int)atoi(buf))  ||  port > 65535) {
         port = CONN_PORT_SMTP;
     }
     if (!ConnNetInfo_GetValue(0, "MX_HOST", buf, sizeof(buf), 0)
